fadding a mapper for user input - adding new error response code - separating validation logic
