+++ app/mappers/users.js
@@ -1,0 +1,4 @@
+exports.formatUserInput = payload => {
+  const { first_name, last_name, email, password } = payload;
+  return { firstName: first_name, lastName: last_name, email, password };
+};
