+++ app/middlewares/errors.js
@@ -10,7 +10,7 @@ const statusCodes = {
   [errors.BAD_REQUEST_ERROR]: 400,
   [errors.NOT_FOUND_ERROR]: 404,
   [errors.CONFLICT_ERROR]: 409,
-  [errors.SCHEMA_ERROR]: 400
+  [errors.SCHEMA_ERROR]: 422
 };
 
 exports.handle = (error, req, res, next) => {
