+++ app/routes.js
@@ -1,7 +1,8 @@
 const { healthCheck } = require('./controllers/healthCheck');
 const { signUp } = require('./controllers/users');
 const { userSchema } = require('./schemas/users');
-const { validateBySchema, checkExistingEmail } = require('./middlewares/validators/users');
+const { checkExistingEmail } = require('./middlewares/validators/users');
+const { validateBySchema } = require('./middlewares/validators/schemaValidators');
 
 exports.init = app => {
   app.get('/health', healthCheck);
