+++ app/schemas/users.js
@@ -1,15 +1,17 @@
 const emailRegex = /^[a-zA-Z]+\.[a-zA-Z]+@wolox.+((co)|(ar)|(mx))$/;
 
 exports.userSchema = {
-  firstName: {
+  first_name: {
     in: ['body'],
+    field: 'firstName',
     isEmpty: {
       negated: true,
       errorMessage: 'firstName is required'
     }
   },
-  lastName: {
+  last_name: {
     in: ['body'],
+    field: 'lastName',
     isEmpty: {
       negated: true,
       errorMessage: 'lastName is required'
